handling custom error

+++ src/controllers/user.controller.js
@@ -1,14 +1,22 @@
+import { ForbiddenHandler } from "../middlewares/errorHandler.js";
 import { User } from "../models/user.models.js";
 
 export const getAllUsers = async (req, res) => {
   try {
     const user = await User.find();
 
+    if (!user) {
+      throw new ForbiddenHandler("all users cannot be found!");
+    }
+
     res.status(200).json({
       message: "all user get request",
       data: user,
     });
   } catch (error) {
+    if (error instanceof Error) {
+      error;
+    }
     next(error);
   }
 };
@@ -16,6 +24,20 @@ export const getAllUsers = async (req, res) => {
 export const registerUser = async (req, res) => {
   try {
     const { firstName, lastName, email, password } = req.body;
+
+    if (!firstName) {
+      throw new ForbiddenHandler("FirstName should be there");
+    }
+    if (!lastName) {
+      throw new ForbiddenHandler("lastName should be there");
+    }
+    if (!email) {
+      throw new ForbiddenHandler("email should be there");
+    }
+    if (!password) {
+      throw new ForbiddenHandler("password should be there");
+    }
+
     const user = await User.insertOne({ firstName, lastName, email, password });
 
     res.status(201).json({
@@ -23,6 +45,9 @@ export const registerUser = async (req, res) => {
       data: user,
     });
   } catch (error) {
+    if (error instanceof Error) {
+      error;
+    }
     next(error);
   }
 };
@@ -31,7 +56,9 @@ export const updateUser = async (req, res) => {
   try {
     const { id } = req.params;
     const updateData = req.body;
-
+    if (!id) {
+      throw new ForbiddenHandler("id should be there");
+    }
     const user = await User.updateOne({ _id: id }, { $set: updateData });
 
     res.status(200).json({
@@ -39,6 +66,9 @@ export const updateUser = async (req, res) => {
       data: user,
     });
   } catch (error) {
+    if (error instanceof Error) {
+      error;
+    }
     next(error);
   }
 };
@@ -47,6 +77,9 @@ export const getUserById = async (req, res) => {
   try {
     const { id } = req.params;
 
+    if (!id) {
+      throw new ForbiddenHandler("id should be there");
+    }
     const user = await User.findById(id);
 
     res.status(200).json({
@@ -54,6 +87,9 @@ export const getUserById = async (req, res) => {
       data: user,
     });
   } catch (error) {
+    if (error instanceof Error) {
+      error;
+    }
     next(error);
   }
 };
@@ -61,7 +97,9 @@ export const getUserById = async (req, res) => {
 export const removeUser = async (req, res) => {
   try {
     const { id } = req.params;
-
+    if (!id) {
+      throw new ForbiddenHandler("id should be there");
+    }
     const user = await User.deleteOne({ _id: id });
 
     res.status(200).json({
@@ -69,6 +107,9 @@ export const removeUser = async (req, res) => {
       data: user,
     });
   } catch (error) {
+    if (error instanceof Error) {
+      error;
+    }
     next(error);
   }
 };
